+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add some additional safety checks to the Essentials integration.
  - WIP: Add double-click-to-reorganize wand inventory functionality.
  - Fix the farming wand, add additional safety for unusable materials.
+ - Allow tesseract to pass through glass
 
 ## 2.8.0